ab_networking

Struct Config

pub struct Config {

    pub keypair: Keypair,
    pub listen_on: Vec<Multiaddr>,
    pub listen_on_fallback_to_random_port: bool,
    pub timeout: Duration,
    pub identify: Config,
    pub kademlia: Config,
    pub gossipsub: Option<Config>,
    pub yamux_config: Config,
    pub allow_non_global_addresses_in_dht: bool,
    pub initial_random_query_interval: Duration,
    pub known_peers_registry: Box<dyn KnownPeersRegistry>,
    pub request_response_protocols: Vec<Box<dyn RequestHandler>>,
    pub reserved_peers: Vec<Multiaddr>,
    pub max_established_incoming_connections: u32,
    pub max_established_outgoing_connections: u32,
    pub max_pending_incoming_connections: u32,
    pub max_pending_outgoing_connections: u32,
    pub temporary_bans_cache_size: u32,
    pub temporary_ban_backoff: ExponentialBuilder,
    pub libp2p_metrics: Option<Metrics>,
    pub metrics: Option<SubspaceMetrics>,
    pub protocol_version: String,
    pub bootstrap_addresses: Vec<Multiaddr>,
    pub kademlia_mode: KademliaMode,
    pub external_addresses: Vec<Multiaddr>,

}

Node configuration.

Fields
keypair: Keypair

Identity keypair of a node used for authenticated connections.

listen_on: Vec<Multiaddr>

List of [Multiaddr] on which to listen for incoming connections.

listen_on_fallback_to_random_port: bool

Fallback to random port if specified (or default) port is already occupied.

timeout: Duration

Adds a timeout to the setup and protocol upgrade process for all inbound and outbound connections established through the transport.

identify: Config

The configuration for the Identify behaviour.

kademlia: Config

The configuration for the Kademlia behaviour.

gossipsub: Option<Config>

The configuration for the Gossip behaviour.

yamux_config: Config

Yamux multiplexing configuration.

allow_non_global_addresses_in_dht: bool

Should non-global addresses be added to the DHT?

initial_random_query_interval: Duration

How frequently should random queries be done using Kademlia DHT to populate routing table.

known_peers_registry: Box<dyn KnownPeersRegistry>

A reference to the NetworkingParametersRegistry implementation.

request_response_protocols: Vec<Box<dyn RequestHandler>>

The configuration for the RequestResponsesBehaviour protocol.

reserved_peers: Vec<Multiaddr>

Defines set of peers with a permanent connection (and reconnection if necessary).

max_established_incoming_connections: u32

Established incoming swarm connection limit.

max_established_outgoing_connections: u32

Established outgoing swarm connection limit.

max_pending_incoming_connections: u32

Pending incoming swarm connection limit.

max_pending_outgoing_connections: u32

Pending outgoing swarm connection limit.

temporary_bans_cache_size: u32

How many temporarily banned unreachable peers to keep in memory.

temporary_ban_backoff: ExponentialBuilder

Backoff policy for temporary banning of unreachable peers.

libp2p_metrics: Option<Metrics>

Optional libp2p prometheus metrics. None will disable metrics gathering.

metrics: Option<SubspaceMetrics>

Internal prometheus metrics. None will disable metrics gathering.

protocol_version: String

Defines protocol version for the network peers. Affects network partition.

bootstrap_addresses: Vec<Multiaddr>

Addresses to bootstrap Kademlia network

kademlia_mode: KademliaMode

Kademlia mode. The default value is set to Static(Client). The peer won’t add its address to other peers` Kademlia routing table. Changing this behaviour implies that a peer can provide pieces to others.

external_addresses: Vec<Multiaddr>

Known external addresses to the local peer. The addresses will be added on the swarm start and enable peer to notify others about its reachable address.

Implementations

impl Config

pub fn new( protocol_version: String, keypair: Keypair, prometheus_registry: Option<&mut Registry>, ) -> Self

Creates a new Config. Applies a subspace-specific version prefix to the protocol_version.